[#114] refactor: b2c 모듈에도 사용해야하기 때문에 common으로 이동
