+++ common/src/main/java/itcast/mail/dto/request/MailContent.java
