+++ common/src/main/java/itcast/mail/dto/request/SendMailRequest.java
@@ -1,0 +1,21 @@
+package itcast.mail.dto.request;
+
+import jakarta.validation.constraints.Email;
+import jakarta.validation.constraints.NotBlank;
+import java.util.List;
+import lombok.Builder;
+
+@Builder
+public record SendMailRequest(
+        @Email
+        List<String> receivers,
+        @NotBlank
+        List<MailContent> contents
+) {
+    public static SendMailRequest of(List<String> receivers, List<MailContent> contents) {
+        return SendMailRequest.builder()
+                .receivers(receivers)
+                .contents(contents)
+                .build();
+    }
+}
