+++ common/src/main/java/itcast/mail/sender/EmailSender.java
@@ -1,10 +1,11 @@
-package itcast.mail.dto.request;
+package itcast.mail.sender;
 
 import com.amazonaws.services.simpleemail.model.Body;
 import com.amazonaws.services.simpleemail.model.Content;
 import com.amazonaws.services.simpleemail.model.Destination;
 import com.amazonaws.services.simpleemail.model.Message;
 import com.amazonaws.services.simpleemail.model.SendEmailRequest;
+import itcast.mail.dto.request.SendMailRequest;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.annotation.Value;
