+++ common/src/main/resources/templates/email-template.html
